refactoring queries 8 and 10
Implements a new endpoint to retrieve products that haven't had any transactions in the last 5 days.
It involves creating a DTO to encapsulate the required product information and modifying the product repository and service to fetch and return the data in the desired format.

Also, refactors database schema for materialized view and updates indexes for performance.